+++ src/eyconf/generate_yaml.py
@@ -15,6 +15,7 @@ from typing import (
     TYPE_CHECKING,
     Annotated,
     Any,
+    Dict,
     Union,
     get_args,
     get_origin,
@@ -218,7 +219,7 @@ def __field_to_lines(field: Field[Any], field_type: type, indent=0) -> list[Line
 
     # Parse default
     default_missing = True
-    default_value = None
+    default_value: Any = None
     if isinstance(field.default, _MISSING_TYPE):
         if isinstance(field.default_factory, _MISSING_TYPE):
             default_missing = True
@@ -229,13 +230,17 @@ def __field_to_lines(field: Field[Any], field_type: type, indent=0) -> list[Line
         default_value = field.default
         default_missing = False
 
+    if origin in [dict, Dict] and default_missing:
+        default_value = {}
+        default_missing = False
+
     # No default value only allowed if the field is optional
     if default_missing and not is_optional:
         raise ValueError(
             f"Field '{field.name}' has no default value! You may set one using direct assignment or a default factory."
         )
 
-    # Lists/Sequences
+    # Default value: Lists/Sequences
     if isinstance(default_value, list):
         if len(default_value) == 0:
             default_value = None
@@ -254,6 +259,42 @@ def __field_to_lines(field: Field[Any], field_type: type, indent=0) -> list[Line
 
                 if __is_dataclass_instance(value):
                     lines.append(SequenceLine("", indent=indent + 1))
+                    lines += _dataclass_to_lines(value, indent=indent + 2)
+                    continue
+
+                raise NotImplementedError(
+                    f"Field type {field.type} {args} {origin} is not supported."
+                )
+
+            return lines
+
+    # Default value: dicts
+    if isinstance(default_value, dict):
+        if len(default_value) == 0:
+            """
+            Stuff: {}
+            """
+            lines.append(MapLine(name=field.name, default_value=r"{}", indent=indent))
+            return lines
+        else:
+            """
+            Stuff:
+                key1: value1
+                key2: value2
+            """
+            lines.append(SectionLine(field.name, indent=indent))
+            for key, value in default_value.items():
+                if not isinstance(key, str):
+                    raise ValueError("Only string keys are supported in dict types")
+
+                if __is_primitive_instance(value):
+                    lines.append(
+                        MapLine(name=key, default_value=value, indent=indent + 1)
+                    )
+                    continue
+
+                if __is_dataclass_instance(value):
+                    lines.append(MapLine(name=key, default_value="", indent=indent + 1))
                     lines += _dataclass_to_lines(value, indent=indent + 2)
                     continue